k r a b
Added a crab, a squat crustacean with pincers. Depicted as a reddish-orange crab (as cooked) shown from above, with a round, shelled body, shorted, tentacle-like eyes. Crabs are decapod crustaceans of the infraorder Brachyura, which typically have a very short projecting "tail" (abdomen), usually hidden entirely under the thorax. They live in all the world's oceans, in fresh water, and on land, are generally covered with a thick exoskeleton, and have a single pair of pincers. Many other animals with similar names – such as hermit crabs, king crabs, porcelain crabs, horseshoe crabs, stone crabs, and crab lice – are not true crabs, but many have evolved features similar to true crabs through a process known as carcinisation. This is also a tribute to the song that goes "dun dun dun dundun dundun dun dundun dundun dun dun dundun dundundun" that is crab-related and created by a storm of noise.